Refactor Product import to use interface module
Replaced all imports of Product from '@/components/product/product' with '@/components/product/interface' across actions, admin pages, and store files for consistency and improved code organization.
