+++ app/admin/products/create/page.tsx
@@ -2,7 +2,7 @@
 
 import { createProduct } from "@/actions";
 import ProductForm from "@/components/admin/product-form";
-import { Product } from "@/components/product/product";
+import { Product } from "@/components/product/interface";
 import { useState } from "react";
 
 export default function CreateProductPage() {
